feat(backend): realign structure, move to chi
