+++ backend/errors.go
@@ -1,11 +1,2 @@
 package main
 
-import "strings"
-
-type ValidationError struct {
-	Errors []string
-}
-
-func (v ValidationError) Error() string {
-	return strings.Join(v.Errors, ", ")
-}
