+++ backend/mailer/mailer.go
@@ -1,4 +1,4 @@
-package main
+package mailer
 
 import (
 	"bytes"
