+++ backend/ticketing/ticketing_test.go
@@ -1,19 +1,92 @@
-package main_test
+package ticketing_test
 
 import (
 	"context"
 	"crypto/ed25519"
 	"crypto/rand"
 	"errors"
+	"os"
 	"strings"
 	"testing"
 	"time"
 
+	"conf/mailer"
+	"conf/ticketing"
+	"conf/user"
+	"github.com/getsentry/sentry-go"
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/rs/zerolog/log"
 	"gocloud.dev/blob"
-
-	main "conf"
+	_ "gocloud.dev/blob/fileblob"
 )
+
+var database *pgxpool.Pool
+var bucket *blob.Bucket
+var mailSender *mailer.Mailer
+
+func TestMain(m *testing.M) {
+	databaseUrl, ok := os.LookupEnv("DATABASE_URL")
+	if !ok {
+		databaseUrl = "postgres://postgres:password@localhost:5432/conf?sslmode=disable"
+	}
+
+	tempDir, err := os.MkdirTemp(os.TempDir(), "teknologi-umum-conference")
+	if err != nil {
+		log.Fatal().Err(err).Msg("creating temporary directory")
+		return
+	}
+
+	blobUrl, ok := os.LookupEnv("BLOB_URL")
+	if !ok {
+		blobUrl = "file://" + tempDir
+	}
+
+	smtpHostname, ok := os.LookupEnv("SMTP_HOSTNAME")
+	if !ok {
+		smtpHostname = "localhost"
+	}
+	smtpPort, ok := os.LookupEnv("SMTP_PORT")
+	if !ok {
+		smtpPort = "1025"
+	}
+	smtpFrom, ok := os.LookupEnv("SMTP_FROM")
+	if !ok {
+		smtpFrom = ""
+	}
+	smtpPassword, ok := os.LookupEnv("SMTP_PASSWORD")
+	if !ok {
+		smtpPassword = ""
+	}
+
+	_ = sentry.Init(sentry.ClientOptions{})
+
+	database, err = pgxpool.New(context.Background(), databaseUrl)
+	if err != nil {
+		log.Fatal().Err(err).Msg("creating pgx pool instance")
+		return
+	}
+
+	bucket, err = blob.OpenBucket(context.Background(), blobUrl)
+	if err != nil {
+		log.Fatal().Err(err).Msg("creating bucket instance")
+		return
+	}
+
+	mailSender = mailer.NewMailSender(&mailer.MailConfiguration{
+		SmtpHostname: smtpHostname,
+		SmtpPort:     smtpPort,
+		SmtpFrom:     smtpFrom,
+		SmtpPassword: smtpPassword,
+	})
+
+	exitCode := m.Run()
+
+	_ = os.RemoveAll(tempDir)
+	_ = bucket.Close()
+	database.Close()
+
+	os.Exit(exitCode)
+}
 
 func TestNewTicketDomain(t *testing.T) {
 	// Create mock dependencies.
@@ -21,11 +94,11 @@ func TestNewTicketDomain(t *testing.T) {
 	bucket := &blob.Bucket{}
 	privateKey := ed25519.PrivateKey{}
 	publicKey := ed25519.PublicKey{}
-	mailer := &main.Mailer{}
+	mailer := &mailer.Mailer{}
 
 	// Group the tests with t.Run().
 	t.Run("all dependencies set", func(t *testing.T) {
-		ticketDomain, err := main.NewTicketDomain(db, bucket, privateKey, publicKey, mailer)
+		ticketDomain, err := ticketing.NewTicketDomain(db, bucket, privateKey, publicKey, mailer)
 		if err != nil {
 			t.Errorf("NewTicketDomain failed: %v", err)
 		}
@@ -35,7 +108,7 @@ func TestNewTicketDomain(t *testing.T) {
 	})
 
 	t.Run("nil database", func(t *testing.T) {
-		ticketDomain, err := main.NewTicketDomain(nil, bucket, privateKey, publicKey, mailer)
+		ticketDomain, err := ticketing.NewTicketDomain(nil, bucket, privateKey, publicKey, mailer)
 		if err == nil {
 			t.Error("NewTicketDomain did not return error with nil database")
 		}
@@ -45,7 +118,7 @@ func TestNewTicketDomain(t *testing.T) {
 	})
 
 	t.Run("nil bucket", func(t *testing.T) {
-		ticketDomain, err := main.NewTicketDomain(db, nil, privateKey, publicKey, mailer)
+		ticketDomain, err := ticketing.NewTicketDomain(db, nil, privateKey, publicKey, mailer)
 		if err == nil {
 			t.Error("NewTicketDomain did not return error with nil bucket")
 		}
@@ -55,7 +128,7 @@ func TestNewTicketDomain(t *testing.T) {
 	})
 
 	t.Run("nil private key", func(t *testing.T) {
-		ticketDomain, err := main.NewTicketDomain(db, bucket, nil, publicKey, mailer)
+		ticketDomain, err := ticketing.NewTicketDomain(db, bucket, nil, publicKey, mailer)
 		if err == nil {
 			t.Error("NewTicketDomain did not return error with nil private key")
 		}
@@ -65,7 +138,7 @@ func TestNewTicketDomain(t *testing.T) {
 	})
 
 	t.Run("nil public key", func(t *testing.T) {
-		ticketDomain, err := main.NewTicketDomain(db, bucket, privateKey, nil, mailer)
+		ticketDomain, err := ticketing.NewTicketDomain(db, bucket, privateKey, nil, mailer)
 		if err == nil {
 			t.Error("NewTicketDomain did not return error with nil public key")
 		}
@@ -75,7 +148,7 @@ func TestNewTicketDomain(t *testing.T) {
 	})
 
 	t.Run("nil mailer", func(t *testing.T) {
-		ticketDomain, err := main.NewTicketDomain(db, bucket, privateKey, publicKey, nil)
+		ticketDomain, err := ticketing.NewTicketDomain(db, bucket, privateKey, publicKey, nil)
 		if err == nil {
 			t.Error("NewTicketDomain did not return error with nil mailer")
 		}
@@ -92,12 +165,12 @@ func TestTicketDomain_StorePaymentReceipt(t *testing.T) {
 		return
 	}
 
-	ticketDomain, err := main.NewTicketDomain(database, bucket, privateKey, publicKey, mailSender)
+	ticketDomain, err := ticketing.NewTicketDomain(database, bucket, privateKey, publicKey, mailSender)
 	if err != nil {
 		t.Fatalf("creating a ticket domain instance: %s", err.Error())
 	}
 
-	userDomain := main.NewUserDomain(database)
+	userDomain := user.NewUserDomain(database)
 
 	t.Run("Invalid Email and photo", func(t *testing.T) {
 		err := ticketDomain.StorePaymentReceipt(context.Background(), "", nil, "")
@@ -105,7 +178,7 @@ func TestTicketDomain_StorePaymentReceipt(t *testing.T) {
 			t.Error("expecting an error, got nil instead")
 		}
 
-		var validationError *main.ValidationError
+		var validationError *ticketing.ValidationError
 		if errors.As(err, &validationError) {
 			if len(validationError.Errors) != 3 {
 				t.Errorf("expecting three errors, got %d", len(validationError.Errors))
@@ -118,7 +191,7 @@ func TestTicketDomain_StorePaymentReceipt(t *testing.T) {
 		defer cancel()
 
 		email := "johndoe+happy@example.com"
-		err := userDomain.CreateParticipant(ctx, main.CreateParticipantRequest{
+		err := userDomain.CreateParticipant(ctx, user.CreateParticipantRequest{
 			Name:  "John Doe",
 			Email: email,
 		})
@@ -137,7 +210,7 @@ func TestTicketDomain_StorePaymentReceipt(t *testing.T) {
 		defer cancel()
 
 		email := "johndoe+happy@example.com"
-		err := userDomain.CreateParticipant(ctx, main.CreateParticipantRequest{
+		err := userDomain.CreateParticipant(ctx, user.CreateParticipantRequest{
 			Name:  "John Doe",
 			Email: email,
 		})
@@ -167,7 +240,7 @@ func TestTicketDomain_StorePaymentReceipt(t *testing.T) {
 			t.Error("expecting an error, got nil instead")
 		}
 
-		if err != nil && !errors.Is(err, main.ErrUserEmailNotFound) {
+		if err != nil && !errors.Is(err, ticketing.ErrUserEmailNotFound) {
 			t.Errorf("unexpected error: %s", err.Error())
 		}
 	})
@@ -180,12 +253,12 @@ func TestTicketDomain_ValidatePaymentReceipt(t *testing.T) {
 		return
 	}
 
-	ticketDomain, err := main.NewTicketDomain(database, bucket, privateKey, publicKey, mailSender)
+	ticketDomain, err := ticketing.NewTicketDomain(database, bucket, privateKey, publicKey, mailSender)
 	if err != nil {
 		t.Fatalf("creating a ticket domain instance: %s", err.Error())
 	}
 
-	userDomain := main.NewUserDomain(database)
+	userDomain := user.NewUserDomain(database)
 
 	t.Run("Invalid email", func(t *testing.T) {
 		_, err := ticketDomain.ValidatePaymentReceipt(context.Background(), "")
@@ -193,7 +266,7 @@ func TestTicketDomain_ValidatePaymentReceipt(t *testing.T) {
 			t.Error("expecting an error, got nil instead")
 		}
 
-		var validationError *main.ValidationError
+		var validationError *ticketing.ValidationError
 		if errors.As(err, &validationError) {
 			if len(validationError.Errors) != 1 {
 				t.Errorf("expecting one error, got %d", len(validationError.Errors))
@@ -210,7 +283,7 @@ func TestTicketDomain_ValidatePaymentReceipt(t *testing.T) {
 			t.Error("expecting an error, got nil")
 		}
 
-		if !errors.Is(err, main.ErrInvalidTicket) {
+		if !errors.Is(err, ticketing.ErrInvalidTicket) {
 			t.Errorf("expecting an error of ErrInvalidTicket, instead got %s", err.Error())
 		}
 	})
@@ -220,7 +293,7 @@ func TestTicketDomain_ValidatePaymentReceipt(t *testing.T) {
 		defer cancel()
 
 		email := "johndoe+happy@example.com"
-		err := userDomain.CreateParticipant(ctx, main.CreateParticipantRequest{
+		err := userDomain.CreateParticipant(ctx, user.CreateParticipantRequest{
 			Name:  "John Doe",
 			Email: email,
 		})
@@ -250,7 +323,7 @@ func TestTicketDomain_VerifyIsStudent(t *testing.T) {
 		t.Fatalf("generating new ed25519 key: %s", err.Error())
 		return
 	}
-	ticketDomain, err := main.NewTicketDomain(database, bucket, privateKey, publicKey, mailSender)
+	ticketDomain, err := ticketing.NewTicketDomain(database, bucket, privateKey, publicKey, mailSender)
 	if err != nil {
 		t.Fatalf("creating a ticket domain instance: %s", err.Error())
 	}
@@ -260,7 +333,7 @@ func TestTicketDomain_VerifyIsStudent(t *testing.T) {
 			t.Error("expecting an error, got nil instead")
 		}
 
-		var validationError *main.ValidationError
+		var validationError *ticketing.ValidationError
 		if errors.As(err, &validationError) {
 			if len(validationError.Errors) != 1 {
 				t.Errorf("expecting one error, got %d", len(validationError.Errors))
